+++ src/app/views/correct_questions/create.js.erb
@@ -1,0 +1,1 @@
+$("#js-multiple-choice-answer-for-quiz-<%= @multiple_choice_question.id %>").replaceWith("<%= j(render('multiple_choice_questions/templates/correct_question', multiple_choice_question: @multiple_choice_question)) %>");
