+++ src/app/views/multiple_choice_questions/finish.js.erb
@@ -1,0 +1,1 @@
+$("#multiple-choice-result").after("<%= j(render('multiple_choice_questions/templates/result', { multiple_choice_question: @multiple_choice_question, multiple_choice_challenger: @multiple_choice_challenger, next_part: @next_part })) %>");
