+++ src/app/views/multiple_choice_questions/templates/_incorrect_question.html.erb
@@ -1,0 +1,51 @@
+<div id="multiple-choice-answer">
+  <div class="incorrect-answer">
+    <svg class="h-8 w-8 text-red-500"  width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+      <path stroke="none" d="M0 0h24v24H0z"/>
+      <line x1="18" y1="6" x2="6" y2="18" /><line x1="6" y1="6" x2="18" y2="18" />
+    </svg>
+  </div>
+  <div class="grid grid-cols-1 place-items-center mb-4">
+    <p class="text-5xl text-center" id="answer-body"><%= multiple_choice_question.body %></p>
+    <div id="sound" class="listening-sound">
+      <svg class="h-16 w-16 text-center text-gray-500"  fill="none" viewBox="0 0 24 24" stroke="currentColor">
+        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15.536 8.464a5 5 0 010 7.072m2.828-9.9a9 9 0 010 12.728M5.586 15H4a1 1 0 01-1-1v-4a1 1 0 011-1h1.586l4.707-4.707C10.923 3.663 12 4.109 12 5v14c0 .891-1.077 1.337-1.707.707L5.586 15z"/>
+      </svg>
+    </div>
+  </div>
+  <div id="js-choice-button-for-quiz-<%= multiple_choice_question.id %>">
+    <div class="mx-auto grid gap-6 sm:w-5/12 sm:grid-cols-2 md:w-5/12 md:grid-cols-2 lg:w-10/12 lg:grid-cols-2">
+      <% multiple_choice_question.multiple_choices.shuffle.each do |choice| %>
+        <% if choice.is_answer %>
+          <div class="bg-green-200 rounded-2xl shadow-xl px-3 py-3 sm:px-3 lg:px-3">
+            <div class="mb-2 space-y-4">
+              <h3 class="text-xl font-semibold text-center text-purple-900" id="correct-answer-body"><%= choice.body %></h3>
+              <div id="correct-body-sound" class="listening-sound">
+                <svg class="h-16 w-16 text-center text-gray-500"  fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15.536 8.464a5 5 0 010 7.072m2.828-9.9a9 9 0 010 12.728M5.586 15H4a1 1 0 01-1-1v-4a1 1 0 011-1h1.586l4.707-4.707C10.923 3.663 12 4.109 12 5v14c0 .891-1.077 1.337-1.707.707L5.586 15z"/>
+                </svg>
+              </div>
+              <div>
+                <%= image_tag choice.image.url %>
+              </div>
+            </div>
+          </div>
+        <% else %>
+          <div class="bg-red-200 rounded-2xl shadow-xl px-3 py-3 sm:px-3 lg:px-3">
+            <div class="mb-2 space-y-4">
+              <h3 class="text-xl font-semibold text-center text-purple-900" id="incorrect-answer-body"><%= choice.body %> </h3>
+              <div id="incorrect-body-sound" class="listening-sound">
+                <svg class="h-16 w-16 text-center text-gray-500"  fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15.536 8.464a5 5 0 010 7.072m2.828-9.9a9 9 0 010 12.728M5.586 15H4a1 1 0 01-1-1v-4a1 1 0 011-1h1.586l4.707-4.707C10.923 3.663 12 4.109 12 5v14c0 .891-1.077 1.337-1.707.707L5.586 15z"/>
+                </svg>
+              </div>
+              <div>
+                <%= image_tag choice.image.url %>
+              </div>
+            </div>
+          </div>
+        <% end %>
+      <% end %>
+    </div>
+  </div>
+</div>
