+++ src/app/views/multiple_choice_questions/templates/_result.html.erb
@@ -1,0 +1,13 @@
+<div>
+  <p >あなたの得点は、<%=@multiple_choice_questions.count%>点中<%=@multiple_choice_challenger.score%>点でした！</p>
+  <div class="to-tweet">
+    <div class="w-full px-3 py-4 bg-blue-300 rounded-md hover:bg-blue-400 focus:outline-none duration-100 ease-in-out">
+      <svg class="h-8 w-8 text-blue-500"viewBox="0 0 24 24"fill="none"stroke="currentColor"stroke-width="2"stroke-linecap="round"stroke-linejoin="round">
+        <path d="M23 3a10.9 10.9 0 0 1-3.14 1.53 4.48 4.48 0 0 0-7.86 3v1A10.66 10.66 0 0 1 3 4s-4 9 5 13a11.64 11.64 0 0 1-7 2c9 5 20 0 20-11.5a4.5 4.5 0 0 0-.08-.83A7.72 7.72 0 0 0 23 3z" />
+      </svg>
+      <p class="text-white">"ツイッターで結果を共有する！"</p>
+    </div>
+  </div>
+  <%= link_to 'トップページへ！', dashboard_path %>
+  <%= link_to '次のレベルへ！', level_part_path(next_part) %>
+</div>
