+++ src/app/views/multiple_choice_questions/top.html.erb
@@ -1,0 +1,6 @@
+
+<div>
+  <% @level_parts.each do |level_part| %>
+    <%= link_to "#{level_part.level}-#{level_part.part}", level_part_multiple_choice_start_path(level_part) %>
+  <% end %>
+</div>
