+++ src/app/views/shared/_after_login_header.html.erb
@@ -28,7 +28,7 @@
             <a class="inline-block py-2 px-4 text-black font-bold no-underline" href="#">Phonics Trainerの使い方</a>
           </li>
           <li class="mr-3">
-            <a class="inline-block text-black no-underline hover:text-gray-800 hover:text-underline py-2 px-4" href="#">Training一覧</a>
+            <%= link_to "ゲーム一覧へ", multiple_choice_questions_path, class:"inline-block text-black no-underline hover:text-gray-800 hover:text-underline py-2 px-4" %>
           </li>
           <li class="mr-3">
             <a class="inline-block text-black no-underline hover:text-gray-800 hover:text-underline py-2 px-4" href="#">Lessonと動画一覧</a>
