+++ src/config/webpack/environment.js
@@ -1,5 +1,13 @@
 const { environment } = require('@rails/webpacker')
 
+const webpack = require('webpack')
+environment.plugins.prepend('Provide',
+    new webpack.ProvidePlugin({
+        $: 'jquery/src/jquery',
+        jQuery: 'jquery/src/jquery'
+    })
+)
+
 module.exports = environment
 
 function hotfixPostcssLoaderConfig (subloader) {
